requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
discord==2.3.2
discord.py==2.3.2
et-xmlfile==1.1.0
fonttools==4.49.0
frozenlist==1.4.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
rauth==0.7.3
requests==2.31.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tzdata==2023.4
urllib3==2.1.0
yarl==1.9.4
